docs(CodeEditor): rename Editor to CodeEditor
