+++ docs/src/components/CodeEditor/index.js
@@ -1,0 +1,1 @@
+export default, { EDITOR_BACKGROUND_COLOR, EDITOR_GUTTER_COLOR } from './CodeEditorUniveral'
