+++ docs/src/components/ComponentDoc/ComponentExample/ComponentExampleRenderEditor.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { Button, Popup, Segment, Menu } from 'semantic-ui-react'
 
-import Editor, { EDITOR_BACKGROUND_COLOR } from 'docs/src/components/Editor/Editor'
+import Editor, { EDITOR_BACKGROUND_COLOR } from 'docs/src/components/CodeEditor'
 import { updateForKeys } from 'docs/src/hoc'
 import formatCode from 'docs/src/utils/formatCode'
