+++ docs/src/components/ComponentDoc/ComponentExample/ComponentExampleRenderHtml.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { Divider } from 'semantic-ui-react'
 
-import Editor, { EDITOR_BACKGROUND_COLOR } from 'docs/src/components/Editor/Editor'
+import Editor, { EDITOR_BACKGROUND_COLOR } from 'docs/src/components/CodeEditor'
 import formatCode from 'docs/src/utils/formatCode'
 
 const rootStyle = {
