--- docs/src/components/Editor/index.js
@@ -1,1 +1,0 @@
-export default from './Editor'
